+++ features/bootstrap/ApplicationContext.php
@@ -7,6 +7,7 @@ use Behat\Behat\Context\SnippetAcceptingContext;
 use Behat\Gherkin\Node\PyStringNode;
 use PhpZone\PhpZone\Application;
 use Symfony\Component\Console\Tester\ApplicationTester;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
  * Defines application features from the specific context.
@@ -29,7 +30,8 @@ class ApplicationContext implements Context, SnippetAcceptingContext
 
     private function setApplicationTest()
     {
-        $application = new Application('0.1.0');
+        $container = new ContainerBuilder();
+        $application = new Application('x.y.z', $container);
         $application->setAutoExit(false);
         $this->application = $application;